% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSummarizedExperiment.R
\name{addTechRepl}
\alias{addTechRepl}
\title{Add technical replicate information to design}
\usage{
addTechRepl(
  experimentDesign,
  techReplCol,
  subjectCol,
  cond1Name,
  cond2Name = NULL
)
}
\description{
Add technical replicate information to design
}
\keyword{internal}
