% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_name_parsing.R
\name{assembleComparisonConditionMapping}
\alias{assembleComparisonConditionMapping}
\title{This is called in Limma to create a mapping for comparison strings to conditions
used later when writing output.}
\usage{
assembleComparisonConditionMapping(
  conditionComparisonMapping,
  seperator = " - "
)
}
\description{
This is called in Limma to create a mapping for comparison strings to conditions
used later when writing output.
}
