% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxQuantTransform.R
\name{assign_uniprot_ids_mq}
\alias{assign_uniprot_ids_mq}
\title{This function uses regex to get the first uniprot assession and the
corresponding gene and description and adds them to a protein intensity table 
coming from maxquant}
\usage{
assign_uniprot_ids_mq(proteinGroups)
}
\description{
This function uses regex to get the first uniprot assession and the
corresponding gene and description and adds them to a protein intensity table 
coming from maxquant
}
