% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reshape_data.R
\name{computeImputedCounts}
\alias{computeImputedCounts}
\title{Count the number of imputed features by condition and write to wide table}
\usage{
computeImputedCounts(intensityDF)
}
\arguments{
\item{intensityDF}{Long format table with raw and imputed intensities. Each row is a feature (protein/peptide).}
}
\description{
Count the number of imputed features by condition and write to wide table
}
