% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reshape_data.R
\name{computeReplicateCounts}
\alias{computeReplicateCounts}
\title{Count the replicates in each condition and write to wide table}
\usage{
computeReplicateCounts(intensityDF)
}
\arguments{
\item{intensityDF}{Long format table with raw and imputed intensities. Each row is a feature (protein/peptide).}
}
\description{
Count the replicates in each condition and write to wide table
}
