% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_name_sanitizer.R
\name{condition_name_decode_intensity_data}
\alias{condition_name_decode_intensity_data}
\title{Decode condition names to return initial names provided in input}
\usage{
condition_name_decode_intensity_data(dt, dict)
}
\description{
Decode condition names to return initial names provided in input
}
