% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_name_sanitizer.R
\name{condition_name_encoder}
\alias{condition_name_encoder}
\title{Create sanitized condition names to be valid for input to makeContrasts}
\usage{
condition_name_encoder(dt, condition_col_name)
}
\description{
Create sanitized condition names to be valid for input to makeContrasts
}
