% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructSummarizedExperiment.R
\name{constructSummarizedExperiment}
\alias{constructSummarizedExperiment}
\title{This function creates a summarized experiment object from a protein intensity table and experiment design.}
\usage{
constructSummarizedExperiment(
  experimentDesign,
  proteinIntensities,
  listMetadata
)
}
\arguments{
\item{experimentDesign}{XX.}

\item{proteinIntensities}{XX}
}
\value{
A SummarizedExperiment object
}
\description{
This function creates a summarized experiment object from a protein intensity table and experiment design.
}
