% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxQuantTransform.R
\name{convert_protein_groups_to_universal}
\alias{convert_protein_groups_to_universal}
\title{This is a wrapper for the mq parser functions that can be used to go
from maxquant data to protein intensity tables we can work with}
\usage{
convert_protein_groups_to_universal(proteinGroups)
}
\description{
This is a wrapper for the mq parser functions that can be used to go
from maxquant data to protein intensity tables we can work with
}
