% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDesignMatrix.R
\name{createDesignMatrix}
\alias{createDesignMatrix}
\title{Create the model design matrix}
\usage{
createDesignMatrix(experimentDesign, splineModel = FALSE, df_splines = 5)
}
\arguments{
\item{experimentDesign}{data.frame. Experiment design}

\item{splineModel}{logical. TRUE to prepare an experiment design with splines.}

\item{df_splines}{Number of degrees of freedom to use to fit the spline.}
}
\description{
Create the model design matrix
}
