% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createContrastMatrix.R
\name{createPairwiseComparisons}
\alias{createPairwiseComparisons}
\title{Create pairwise comparisons}
\usage{
createPairwiseComparisons(
  comparisonType = "all",
  condLevels,
  conditionsDict,
  orderConditions = NULL,
  baselineInpuLevel = NULL,
  customComparisons = NULL
)
}
\arguments{
\item{comparisonType}{Type of model / pairwise comparison. One of, "all", "oneVSall", "custom", "spline". 
Default to "all"}

\item{baselineInpuLevel}{vector of baseline level}

\item{contrastLevels}{vector with all condition levels to be considered for building contrasts}

\item{orderCondition}{vector of ordered condition levels}

\item{chosenComparisons}{data.frame of chosen comparisons}
}
\description{
Create pairwise comparisons
}
