% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createResults.R
\name{createResults}
\alias{createResults}
\title{Create final results}
\usage{
createResults(
  IntensityExperiment,
  limmaStats = stats,
  longIntensityDT = longIntensityDT,
  conditionComparisonMapping = conditionComparisonMapping
)
}
\arguments{
\item{IntensityExperiment}{output from `createSummarizedExperiment`}

\item{limmaStats}{Statistic table from limma output}

\item{longIntensityDT}{Long format table with raw and imputed intensities. Each row is a feature (protein/peptide).}

\item{conditionComparisonMapping}{a dataframe matching different limma statistics comparison strings to up and down conditions}
}
\description{
Create final results
}
