% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSummarizedExperiment.R
\name{createSummarizedExperiment}
\alias{createSummarizedExperiment}
\title{This function creates a summarized experiment object from a protein intensity table and experiment design.}
\usage{
createSummarizedExperiment(experimentDesign, proteinIntensities, listMetadata)
}
\arguments{
\item{experimentDesign}{XX.}

\item{proteinIntensities}{XX}
}
\value{
A SummarizedExperiment object
}
\description{
This function creates a summarized experiment object from a protein intensity table and experiment design.
}
