% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{detect_condition_string}
\alias{detect_condition_string}
\title{Use literal string match to find condition name in comparison string}
\usage{
detect_condition_string(conditions, string)
}
\description{
Use literal string match to find condition name in comparison string
}
