% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSummarizedExperiment.R
\name{detectDesignType}
\alias{detectDesignType}
\title{Detect the design type}
\usage{
detectDesignType(
  experimentDesign,
  techReplCol = "TechRepl",
  subjectCol = "Subject"
)
}
\arguments{
\item{experimentDesign}{data.frame. Experiment design provided in input by the user.}

\item{techReplCol}{Name of Technical Replicate column name. Deafult `TechRepl`}

\item{subjectCol}{Name of Subject column name. Deafult `Subject`}
}
\description{
Detect the design type
}
