% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diann_data}
\alias{diann_data}
\title{DIA-NN sample output}
\format{
DIA-INN protein input intensities and experimental design. The protein input data was
obtained applying `diann::diann_maxlfq('diann_report.tsv')` where `diann_report.tsv` is part of the 
`diann` `R` package.
}
\usage{
diann_data
}
\description{
DIA-NN sample output
}
\keyword{datasets}
