% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{fill_out_missing_columns}
\alias{fill_out_missing_columns}
\title{This function ensures that all the missing columns are filled with NA's.
It should probably flag important columns that are missing in the future.}
\usage{
fill_out_missing_columns(comparisonStatistics)
}
\description{
This function ensures that all the missing columns are filled with NA's.
It should probably flag important columns that are missing in the future.
}
