% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxQuantTransform.R
\name{filter_columns_mq}
\alias{filter_columns_mq}
\title{This function removes bad rows in a maxquant protein intensity object}
\usage{
filter_columns_mq(protein.intensities)
}
\description{
This function removes bad rows in a maxquant protein intensity object
}
