% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{filter_stats_table_on_comparison}
\alias{filter_stats_table_on_comparison}
\title{This function takes the statistics table from limma and retrieves a single comparison's worth of statistics.}
\usage{
filter_stats_table_on_comparison(statisticsTable, comparison)
}
\description{
This function takes the statistics table from limma and retrieves a single comparison's worth of statistics.
}
