% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareEnrichment.R
\name{filterComparisonRows}
\alias{filterComparisonRows}
\title{This function filter out rows in the assay that have na statistics 
usually caused by the missing value filter for protein}
\usage{
filterComparisonRows(comparisonIntensityExperiment)
}
\description{
This function filter out rows in the assay that have na statistics 
usually caused by the missing value filter for protein
}
