% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitModelOneCondition.R
\name{fitModelOneCondition}
\alias{fitModelOneCondition}
\title{This function orchestrates imputation, normalization and the binary limma statistics accross all experimental comparisons}
\usage{
fitModelOneCondition(
  longIntensityDT,
  metadataExperiment,
  comparisonType = "all",
  orderCondition = NULL,
  baselineInpuLevel = NULL,
  customComparisonsCond = NULL,
  conditionsDict,
  fitSeparateModels,
  returnDecideTestColumn,
  conditionSeparator
)
}
\arguments{
\item{longIntensityDT}{Output from `preProcess`}

\item{metadataExperiment}{Output from `preProcess`}

\item{comparisonType}{Type of model / pairwise comparison. One of, "all", "oneVSall", "custom", "spline". 
Default to "all"}

\item{conditionsDict}{Condtion dictionary.}

\item{fitSeparateModels}{logical. TRUE to fit separate limma models for each pairwise comparisons 
(e.g. filtering and `lmFit` are run separately by comparison).}

\item{returnDecideTestColumn}{logical. If TRUE the row data of the `CompleteIntensityExperiment` will contain the output from 
`limma::decideTests`. If FALSE a single model is run for all contrasts.}

\item{conditionSeparator}{string. String used to separate up and down condition in output.}
}
\description{
This function orchestrates imputation, normalization and the binary limma statistics accross all experimental comparisons
}
