% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fragpipe_data}
\alias{fragpipe_data}
\title{Fragpipe LFQ}
\format{
List containing the `intensities` table, the experiment `design` and the `parameters` following the Generic Format.
}
\usage{
fragpipe_data
}
\description{
Fragpipe LFQ
}
\details{
The data can be used to run the entire workflow `runGenericDiscovery`.
}
\references{
PRIDE Project PXD026401 https://www.ebi.ac.uk/pride/archive/projects/PXD026401
}
\keyword{datasets}
