% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_qc_report.R
\name{generate_qc_report}
\alias{generate_qc_report}
\title{Generate QC report}
\usage{
generate_qc_report(listResults, output_folder, format = "html")
}
\arguments{
\item{listResults}{list. List of results from `runGenericDiscovery`}

\item{output_folder}{str. Output folder path}

\item{format}{str. 'pdf' or 'html'}
}
\description{
Generate QC report
}
