% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_qc_report.R
\name{generate_separate_qc_reports}
\alias{generate_separate_qc_reports}
\title{Generate separate QC reports}
\usage{
generate_separate_qc_reports(listResults, output_folder)
}
\arguments{
\item{listResults}{list. List of results from `runGenericDiscovery`}

\item{output_folder}{str. Output folder path}
}
\description{
Generate separate QC reports
}
