% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{get_comparison_strings}
\alias{get_comparison_strings}
\title{This functions works out which comparisons were calculated by limma.
Currently, it calculates all of them but it assumes order dependence so not trivial.}
\usage{
get_comparison_strings(results.quant)
}
\description{
This functions works out which comparisons were calculated by limma.
Currently, it calculates all of them but it assumes order dependence so not trivial.
}
