% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{get_condition_string}
\alias{get_condition_string}
\title{Use literal string match to find condition name in comparison string}
\usage{
get_condition_string(conditions, comparison, position = 1)
}
\description{
Use literal string match to find condition name in comparison string
}
