% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSummarizedExperiment.R
\name{getConditionName}
\alias{getConditionName}
\title{Extract condition names from experiment design}
\usage{
getConditionName(inputDesign)
}
\description{
Extract condition names from experiment design
}
\keyword{internal}
