% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_name_parsing.R
\name{getDownCondition}
\alias{getDownCondition}
\title{This function uses the condition comparison mapping to get the down condition}
\usage{
getDownCondition(conditionComparisonMapping, comparison.string)
}
\description{
This function uses the condition comparison mapping to get the down condition
}
