% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition_name_parsing.R
\name{getUpCondition}
\alias{getUpCondition}
\title{This function uses the condition comparison mapping to get the up condition}
\usage{
getUpCondition(conditionComparisonMapping, comparison.string)
}
\description{
This function uses the condition comparison mapping to get the up condition
}
