% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSummarizedExperiment.R
\name{hasConditionOnly}
\alias{hasConditionOnly}
\title{Check for only one condition in experiment design}
\usage{
hasConditionOnly(inputDesign)
}
\description{
Check for only one condition in experiment design
}
\keyword{internal}
