% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSummarizedExperiment.R
\name{hasConditionWithTimeDose}
\alias{hasConditionWithTimeDose}
\title{Check for only one condition + time/dose in experiment design}
\usage{
hasConditionWithTimeDose(inputDesign)
}
\description{
Check for only one condition + time/dose in experiment design
}
\keyword{internal}
