% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSummarizedExperiment.R
\name{hasEnoughLevelsInConditions}
\alias{hasEnoughLevelsInConditions}
\title{Check for enough levels in conditions in the experiment design}
\usage{
hasEnoughLevelsInConditions(experimentDesign, condName)
}
\description{
Check for enough levels in conditions in the experiment design
}
\keyword{internal}
