% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSummarizedExperiment.R
\name{hasEnoughNumberReplPerGroup}
\alias{hasEnoughNumberReplPerGroup}
\title{Check for enough replicates given the design}
\usage{
hasEnoughNumberReplPerGroup(experimentDesign, cond1Name, cond2Name = NULL)
}
\description{
Check for enough replicates given the design
}
\keyword{internal}
