% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSummarizedExperiment.R
\name{hasRepMeas}
\alias{hasRepMeas}
\title{Check for repeated measurements in design}
\usage{
hasRepMeas(experimentDesign, subjectCol = "Subject", repCondName)
}
\description{
Check for repeated measurements in design
}
\keyword{internal}
