% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{importData}
\alias{importData}
\title{This function creates a summarized experiment object from a protein intensity table and experiment design.}
\usage{
importData(
  experimentDesign,
  proteinIntensities,
  secondConditionName,
  normalisationMethod = NULL,
  species = NULL,
  labellingMethod = NULL
)
}
\arguments{
\item{experimentDesign}{data.frame. Experiment design provided in input by the user. Required columsn are: `SampleName` and `Condition`.}

\item{proteinIntensities}{data.frame. Wide matrix of intensities. Rows are proteins and columns are SampleNames. Required column: `ProteinId`.}

\item{normalisationMethod}{Normalisation method. One of "None" or "Median".}

\item{species}{Species. One of 'Human', 'Mouse', 'Yeast', 'Other'}

\item{labellingMethod}{One of 'LFQ' or 'TMT'}
}
\value{
A SummarizedExperiment object
}
\description{
This function creates a summarized experiment object from a protein intensity table and experiment design.
}
