% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeLFQ.R
\name{imputeLFQ}
\alias{imputeLFQ}
\title{This function does MNAR imputation}
\usage{
imputeLFQ(myQuantDT, id_type, int_type, f_imputeStDev, f_imputePosition)
}
\arguments{
\item{myQuantDT}{quantification data}

\item{id_type}{the id column of myQuantDT that indicates}

\item{int_type}{the column of myQuantDT to be imputed}

\item{f_imputeStDev}{The Standard Deviation parameter for MNAR Imputation}

\item{f_imputePosition}{The Position parameter for MNAR Imputation}
}
\value{
quantification data with missing values imputed
}
\description{
This function does MNAR imputation
}
\examples{
 ## dt_int <- imputeLFQ(myQuantDT = dt_int,
 ## id_type = "id",
 ## int_type = "log2NInt", #log2Intensity
 ## 0.3,
 ## 1.8,
 ## )
}
