% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initialiseLongIntensityDT}
\alias{initialiseLongIntensityDT}
\title{This function performs the log2 conversion for intensities larger than 0 and initialise the imputed column.}
\usage{
initialiseLongIntensityDT(IntensityExperiment)
}
\arguments{
\item{IntensityExperiment}{Output from createSummarizedExperiment}
}
\description{
This function performs the log2 conversion for intensities larger than 0 and initialise the imputed column.
}
