% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaPairwiseComparisons.R
\name{limmaPairwiseComparisons}
\alias{limmaPairwiseComparisons}
\title{This function performs the differential expression analysis with limma including all pairwise comparisons 
using the condition provided}
\usage{
limmaPairwiseComparisons(
  conditionColname,
  longIntensityDT,
  featureIdType,
  intensityType,
  runIdColname,
  repColname,
  returnDecideTestColumn,
  conditionSeparator,
  pairwiseComparisons
)
}
\arguments{
\item{conditionColname}{str. Column name for the condition of interest, e.g. `Condition`. Used to build all pairwise comparisons.}

\item{longIntensityDT}{data.table. long table containing intensities for all proteins and samples.}

\item{featureIdType}{str. Column name of feature ID, e.g. `ProteinId`}

\item{intensityType}{str. Column name containing intensities to use for the differential expression analysis, e.g. `log2NIntNorm`}

\item{runIdColname}{str. Column name for the unique identifier of condition and replicate, e.g. `RunId`}

\item{repColname}{str. Replicate column name. e.g. `Replicate`}

\item{returnDecideTestColumn}{logical. If TRUE the row data of the `CompleteIntensityExperiment` will contain the output from 
`limma::decideTests`.}

\item{conditionSeparator}{string. String used to separate up and down condition in output.}

\item{pairwiseComparisons}{chr. XX In the future this will allow the possibility of including only some pairwise comparisons of interest.}

\item{all.comparisons}{logical. TRUE to run differential expression analysis for all pairwise comparisons.}
}
\description{
This function performs the differential expression analysis with limma including all pairwise comparisons 
using the condition provided
}
