% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_enrichment.R
\name{listComparisonExperiments}
\alias{listComparisonExperiments}
\title{This function creates a list of intensity experiments pertaining to each
comparison calculated by the limma pipeline}
\usage{
listComparisonExperiments(CompleteIntensityExperiment)
}
\arguments{
\item{CompleteIntensityExperiment}{SummarizedExperiment containing output from the differential expession analysis.}
}
\description{
This function creates a list of intensity experiments pertaining to each
comparison calculated by the limma pipeline
}
