% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{makeLongIntensityDF}
\alias{makeLongIntensityDF}
\title{This function converts the assay data in a protein intensity to a long data.table object with 
intensities for each protein Id and intensity column}
\usage{
makeLongIntensityDF(IntensityExperiment)
}
\arguments{
\item{IntensityExperiment}{Output from constructSummarizedExperiment}
}
\description{
This function converts the assay data in a protein intensity to a long data.table object with 
intensities for each protein Id and intensity column
}
