% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{makeRLEBoxplotInteractive}
\alias{makeRLEBoxplotInteractive}
\title{Convert output of `plot_rle_boxplot` to interactive using plotly}
\usage{
makeRLEBoxplotInteractive(fig)
}
\arguments{
\item{fig}{non interactive ggplot object}
}
\description{
Convert output of `plot_rle_boxplot` to interactive using plotly
}
