% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveOutputV2.R
\name{MassExpressionTOProtCountV2}
\alias{MassExpressionTOProtCountV2}
\title{MassExpression output to MassDynamics V2 Protein Counts JSON}
\usage{
MassExpressionTOProtCountV2(
  CompleteIntensityExperiment,
  longIntensityDT,
  outputFolder
)
}
\description{
Save output from MassExpression and MDFlexiComparisons in the V2 formats required for protein counts intensities json.
}
