% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxQuantTransform.R
\name{MaxQuantTranform}
\alias{MaxQuantTranform}
\title{Transform MaxQuant proteinGroups to universal input}
\usage{
MaxQuantTranform(proteinGroups, design)
}
\description{
Transform MaxQuant proteinGroups to universal input
}
