% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{mds_plot_experiment}
\alias{mds_plot_experiment}
\title{Plot first 2 dimensions of Multi-Dimensional Scaling plot}
\usage{
mds_plot_experiment(Experiment, log = FALSE)
}
\arguments{
\item{Experiment}{SummarizedExperiment object}

\item{log}{logical. Whether data should be logged before plotting}
}
\description{
Plot first 2 dimensions of Multi-Dimensional Scaling plot
}
