% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mq_lfq_data}
\alias{mq_lfq_data}
\title{MaxQuant LFQ}
\format{
List containing the `intensities` table, the experiment `design` and the `parameters` following the Generic Format.
}
\usage{
mq_lfq_data
}
\description{
MaxQuant LFQ
}
\details{
The data can be used to run the entire workflow `runGenericDiscovery`. 
Download MaxQuant processed data from: https://app.massdynamics.com/experiments/c6fc6c60-fe65-47cb-bd6d-a021f0ed8720#/results-files
}
\references{
PRIDE Project PXD002057

Creedon, H. et al. Identification of novel pathways linking epithelial-to-mesenchymal transition with resistance to HER2-targeted therapy. Oncotarget 7, 11539–11552 (2016).
}
\keyword{datasets}
