% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mq_silac_data}
\alias{mq_silac_data}
\title{MaxQuant SILAC sample output}
\format{
Sample SILAC dataset parsed from the SILAC dataset published in Tyanova S, 2016.
}
\usage{
mq_silac_data
}
\description{
MaxQuant SILAC sample output
}
\keyword{datasets}
