% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveOutputV2.R
\name{oneGroupReplDataV2}
\alias{oneGroupReplDataV2}
\title{Transform data for one protein from a long format to the nested data structure needed for the Replicates tab.}
\usage{
oneGroupReplDataV2(oneProt, GroupColumnName, GroupLabelType)
}
\arguments{
\item{oneProt}{data.table single protein information stored in long format.
Columns required: `ProteinId`, `GeneName`, `Description`, `log2NInt`, `Condition`,
 `Replicate`, `Imputed`.}
}
\description{
Transform data for one protein from a long format to the nested data structure needed for the Replicates tab.
}
