% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveOutput.R
\name{oneProteinReplData}
\alias{oneProteinReplData}
\title{Transform data for one protein from a long format to the nested data structure needed for the Replicates tab.}
\usage{
oneProteinReplData(oneProt)
}
\arguments{
\item{oneProt}{data.table single protein information stored in long format. 
Columns required: `ProteinId`, `GeneName`, `Description`, `log2NInt`, `Condition`,
 `Replicate`, `Imputed`.}
}
\description{
Transform data for one protein from a long format to the nested data structure needed for the Replicates tab.
}
