% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{pca_plot_experiment}
\alias{pca_plot_experiment}
\title{Plot first 2 dimensions of PCA}
\format{
"pdf" or "html". If html an interactive plot is returned using plotly.
}
\usage{
pca_plot_experiment(Experiment, format = "pdf", log = FALSE)
}
\arguments{
\item{Experiment}{SummarizedExperiment object}

\item{log}{logical. Whether data should be logged before plotting}
}
\description{
Plot first 2 dimensions of PCA
}
