% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pd_iPRG_data}
\alias{pd_iPRG_data}
\title{Proteome Discoverer LFQ iRPG}
\format{
Data from XXX cite iRPG paper
}
\usage{
pd_iPRG_data
}
\description{
Proteome Discoverer LFQ iRPG
}
\keyword{datasets}
