% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{plot_condition_cv_distribution}
\alias{plot_condition_cv_distribution}
\title{CV distributuions of a protein summarized by condition of interest}
\usage{
plot_condition_cv_distribution(Experiment)
}
\arguments{
\item{Experiment}{SummarizedExperiment object}
}
\description{
CV distributuions of a protein summarized by condition of interest
}
