% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{plot_consistent_proteins_by_replicate}
\alias{plot_consistent_proteins_by_replicate}
\title{Lollipop plot of the number of proteins identified across all replicates of a condition}
\usage{
plot_consistent_proteins_by_replicate(
  Experiment,
  assayName = "raw",
  condition_colname = "Condition",
  title = "# Consistently identified proteins by condition"
)
}
\arguments{
\item{Experiment}{SummarizedExperiment object}

\item{assayName}{name of assay to use}

\item{condition_colname}{str. Name of grouping condition.}

\item{title}{str. Plot title}
}
\description{
Lollipop plot of the number of proteins identified across all replicates of a condition
}
