% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{plot_heatmap_missingness}
\alias{plot_heatmap_missingness}
\title{Heatmap of the binary intensity matrix of missing values}
\usage{
plot_heatmap_missingness(Experiment, complete = FALSE)
}
\arguments{
\item{Experiment}{SummarizedExperiment object of raw data. 
Intensities values equal to 0 are considered as missing values.}

\item{complete}{logical. TRUE to use all the matrix of intensities. 
FALSE to include only features with at least one missing value across samples.}
}
\description{
Heatmap of the binary intensity matrix of missing values
}
