% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{plot_log_measurement_boxplot}
\alias{plot_log_measurement_boxplot}
\title{Boxplot of log expression values across samples.}
\usage{
plot_log_measurement_boxplot(Experiment)
}
\arguments{
\item{Experiment}{SummarizedExperiment object of raw intensities (pre log-transformation)}
}
\description{
Boxplot of log expression values across samples.
}
\details{
Zero values are treated as missing values and excluded.
}
