% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{plot_ma}
\alias{plot_ma}
\title{MA plot of results}
\usage{
plot_ma(comparison.statistics)
}
\arguments{
\item{comparison.statistics}{data.frame containing 
`FoldChange`, `AveExpr`, and `ProteinId` columns.}
}
\description{
MA plot of results
}
