% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{plot_measurement_boxplot}
\alias{plot_measurement_boxplot}
\title{Boxplot of log expression values across samples.}
\usage{
plot_measurement_boxplot(Experiment, log = FALSE)
}
\arguments{
\item{Experiment}{SummarizedExperiment object}

\item{use_imputed}{logical}
}
\description{
Boxplot of log expression values across samples.
}
