% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{plot_n_identified_proteins_by_replicate}
\alias{plot_n_identified_proteins_by_replicate}
\title{Lollipop plot of the number of proteins identified by replicate}
\usage{
plot_n_identified_proteins_by_replicate(
  Experiment,
  assayName = "raw",
  title = "# Identified proteins by sample"
)
}
\arguments{
\item{Experiment}{SummarizedExperiment object}

\item{assayName}{name of assay to use}

\item{title}{str. Plot title}
}
\description{
Lollipop plot of the number of proteins identified by replicate
}
