% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{plot_pca_experiment}
\alias{plot_pca_experiment}
\title{Plot first 2 dimensions of PCA}
\format{
"pdf" or "html". If html an interactive plot is returned using plotly.
}
\usage{
plot_pca_experiment(
  Experiment,
  format = "pdf",
  log = FALSE,
  onlyDEProteins = FALSE,
  title = "PCA plot"
)
}
\arguments{
\item{Experiment}{SummarizedExperiment object}

\item{format}{'pdf' or 'html'. Prepare image to be rendered for pdf or html Rmd output}

\item{log}{logical. Whether data should be logged before plotting}

\item{onlyDEProteins}{logical. TRUE if only DE proteins should be considered.}
}
\description{
Plot first 2 dimensions of PCA
}
\details{
#' A protein is defined DE is the adjusted PValue is less than 0.05. 
When multiple comparisons are available, the adjusted PValues of the F-test is used.
}
