% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{plot_protein_counts_by_replicate}
\alias{plot_protein_counts_by_replicate}
\title{Lollipop plot of the number of proteins identified by replicate}
\usage{
plot_protein_counts_by_replicate(Experiment)
}
\arguments{
\item{Experiment}{SummarizedExperiment object}
}
\description{
Lollipop plot of the number of proteins identified by replicate
}
