% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{plot_protein_missingness}
\alias{plot_protein_missingness}
\title{Histogram of the distribution of missingness by protein}
\usage{
plot_protein_missingness(Experiment)
}
\arguments{
\item{Experiment}{SummarizedExperiment object}
}
\description{
Histogram of the distribution of missingness by protein  where missingness is defined as 0 values.
}
