% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{plot_replicate_measured_values}
\alias{plot_replicate_measured_values}
\title{Lollipop plot of missingness by intensity column}
\usage{
plot_replicate_measured_values(
  Experiment,
  assayName = "raw",
  title = "Data completedness by samples using protein measurements"
)
}
\arguments{
\item{Experiment}{SummarizedExperiment object}

\item{assayName}{name of assay to use}

\item{title}{str. Plot title}
}
\description{
Lollipop plot of missingness by intensity column where missingness is defined as a value equal to 0 .
}
