% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{plot_rle_boxplot}
\alias{plot_rle_boxplot}
\title{Boxplot of the distribution of the relative log expression values across samples.}
\usage{
plot_rle_boxplot(Experiment, log = FALSE)
}
\arguments{
\item{Experiment}{SummarizedExperiment object}
}
\description{
Boxplot of the distribution of the relative log expression values across samples.
}
