% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{plot_samples_correlation_matrix}
\alias{plot_samples_correlation_matrix}
\title{Correlation plot of samples using DE proteins}
\usage{
plot_samples_correlation_matrix(
  Experiment,
  onlyDEProteins = FALSE,
  title = "All proteins"
)
}
\arguments{
\item{Experiment}{SummarizedExperiment object of imputed data.}

\item{onlyDEProteins}{logical. TRUE if only DE proteins should be considered.}
}
\description{
Correlation plot of samples using DE proteins
}
\details{
#' A protein is defined DE is the adjusted PValue is less than 0.05. 
When multiple comparisons are available, the adjusted PValues of the F-test is used. 
At least 5 DE proteins are required to produce the correlation plot.
}
