% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot of results}
\usage{
plot_volcano(comparison.statistics)
}
\arguments{
\item{comparison.statistics}{data.frame containing 
`FoldChange`, `PValue`, `AdjustedPValue` and `ProteinId` columns}
}
\description{
Volcano plot of results
}
