% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_prot_int}
\alias{prepare_prot_int}
\title{This function performs the log2 conversion and writes the imputed column}
\usage{
prepare_prot_int(IntensityExperiment)
}
\arguments{
\item{IntensityExperiment}{Output from constructSummarizedExperiment}
}
\description{
This function performs the log2 conversion and writes the imputed column
}
