% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareForFiltering.R
\name{prepareForFiltering}
\alias{prepareForFiltering}
\title{Filter protein IDs that are not present in at least one experiment}
\usage{
prepareForFiltering(longIntensityDT, conditionColname, runIdColname)
}
\arguments{
\item{longIntensityDT}{data.table}

\item{conditionColname}{name of column indicating the condition of interest}

\item{runIdColname}{name of column indicating the protein id}
}
\value{
`list`: filtered data.table and list of proteins
}
\description{
Filter protein IDs that are not present in at least one experiment
}
