% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepareLongIntensityDF}
\alias{prepareLongIntensityDF}
\title{This function performs the log2 conversion and writes the imputed column}
\usage{
prepareLongIntensityDF(IntensityExperiment)
}
\arguments{
\item{IntensityExperiment}{Output from constructSummarizedExperiment}
}
\description{
This function performs the log2 conversion and writes the imputed column
}
