% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IO.R
\name{rename_comparison_statistics_export}
\alias{rename_comparison_statistics_export}
\title{MD's protein viz needs specific attribute names so this function does that renaming}
\usage{
rename_comparison_statistics_export(comparisonStatistics)
}
\description{
MD's protein viz needs specific attribute names so this function does that renaming
}
