% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{replicate_missingness_experiment}
\alias{replicate_missingness_experiment}
\title{this function creates a lollipop plot of missingness by intensity column where missingness is defined as a value equal to 0 .}
\usage{
replicate_missingness_experiment(Experiment)
}
\arguments{
\item{Experiment}{SummarizedExperiment object}
}
\description{
this function creates a lollipop plot of missingness by intensity column where missingness is defined as a value equal to 0 .
}
