% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGenericDiscovery.R
\name{runGenericDiscovery}
\alias{runGenericDiscovery}
\title{This function orchestrates the MassExpression workflow (could be called by a  workflow step)}
\usage{
runGenericDiscovery(experimentDesign, proteinIntensities)
}
\arguments{
\item{experimentDesign}{XX.}

\item{proteinIntensities}{XX}
}
\description{
This function orchestrates the MassExpression workflow (could be called by a  workflow step)
}
