% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runOrderedConditionDiscovery.R
\name{runOrderedConditionDiscovery}
\alias{runOrderedConditionDiscovery}
\title{This function orchestrates the MassExpression workflow (could be called by a  workflow step)}
\usage{
runOrderedConditionDiscovery(
  experimentDesign,
  proteinIntensities,
  comparisonType = "all",
  orderConditionsList = NULL,
  baselineConditionList = NULL,
  customComparisonsList = NULL,
  fitSeparateModels = TRUE,
  returnDecideTestColumn = FALSE,
  conditionSeparator = " - "
)
}
\arguments{
\item{experimentDesign}{data.frame. Experiment design provided in input by the user. Required columsn are: `SampleName` and `Condition`.}

\item{proteinIntensities}{data.frame. Wide matrix of intensities. Rows are proteins and columns are SampleNames. Required column: `ProteinId`.}

\item{comparisonType}{Type of model / pairwise comparison. One of, "all", "oneVSall", "custom", "spline". 
Default to "all"}

\item{orderConditionsList}{list. Each entry of the list provides the ordering of the relative condition. 
E.g. list(Condition = c("2h", "4h", "6h)), in this case "2h" is considered the first level of the condition.}

\item{baselineConditionList}{list. Each entry of the list provides the baseline of the relative condition.
list(Condition = c("2h")). This is used to perform the `oneVsall` comparison type. If `NULL`, either the first 
entry of `orderConditions` is used or an automatic detection is performed (imperfect solution).}

\item{customComparisonsList}{list of data.frames. When `comparisonType` is `custom` this field is required 
to establish the custom levels to compare for each condition.}

\item{fitSeparateModels}{logical. TRUE to fit separate limma models for each pairwise comparisons 
(e.g. filtering and `lmFit` are run separately by comparison).}

\item{returnDecideTestColumn}{logical. If TRUE the row data of the `CompleteIntensityExperiment` will contain the output from 
`limma::decideTests`. If FALSE a single model is run for all contrasts.}

\item{conditionSeparator}{string. String used to separate up and down condition in output.}
}
\value{
List of two SummarisedExperiment objects: `IntensityExperiment` 
containing the raw intensities and  `CompleteIntensityExperiment` including 
imputed intensities and the results of the limma DE analysis.
}
\description{
This function orchestrates the MassExpression workflow (could be called by a  workflow step)
}
\details{
Options for `comparisonType` arguments. If `comparisonType = "all"`, `orderConditionsList`/`baselineConditionList` 
and `customComparisonsList` are not considered. If `comparisonType = "custom"` then `customComparisonsList` is required. 
If `comparisonType = "oneVSall"`, `orderConditionsList`/`baselineConditionList` can be provided to establish the order 
of the comparisons. If they are provided, one level of the condition is automatically selected based on automatic ordering
selection.
}
