% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sanitize_strings.R
\name{sanitize_strings}
\alias{sanitize_strings}
\title{Sanitize strings for export}
\usage{
sanitize_strings(any_string)
}
\arguments{
\item{any_string}{string}
}
\description{
Remove invisible characters from string and substitute Unicode characters with empty space
}
