% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_sanitize_strings.R
\name{sanitize_strings_in_dataframe}
\alias{sanitize_strings_in_dataframe}
\title{Sanitize strings in dataframe}
\usage{
sanitize_strings_in_dataframe(df)
}
\arguments{
\item{df}{data frame}
}
\description{
Sanitize all character entries in dataframe. sanitize from invisible and unicode characters.
}
