% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{select_features_for_pca}
\alias{select_features_for_pca}
\title{Subset features to be used for PCA}
\usage{
select_features_for_pca(Experiment, auto_select_features = NULL)
}
\arguments{
\item{Experiment}{SummarizedExperiment object.}

\item{auto_select_features}{str. One of 'de' or empty string.}
}
\description{
Subset features to be used for PCA
}
