% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SEToLongDT}
\alias{SEToLongDT}
\title{This function converts the assay data in a protein intensity to a long data.table object with 
intensities for each protein Id and intensity column}
\usage{
SEToLongDT(IntensityExperiment, assayName = "raw")
}
\arguments{
\item{IntensityExperiment}{Output from createSummarizedExperiment}
}
\description{
This function converts the assay data in a protein intensity to a long data.table object with 
intensities for each protein Id and intensity column
}
