% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareEnrichment.R
\name{validateComparison}
\alias{validateComparison}
\title{This is a simple check to validate the intensity/cls vector match the statistics}
\usage{
validateComparison(rse, comparison)
}
\description{
This is a simple check to validate the intensity/cls vector match the statistics
}
