% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveOutput.R
\name{writeLimmaStatisticsTable}
\alias{writeLimmaStatisticsTable}
\title{Write limma statistics for users to read}
\usage{
writeLimmaStatisticsTable(CompleteIntensityExperiment, outputFolder)
}
\arguments{
\item{CompleteIntensityExperiment}{produced by runLimmaPipeline}

\item{outputFolder}{A path to a folder to write the statistics to.}
}
\description{
Write limma statistics for users to read
}
