% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveOutput.R
\name{writeOutputFile}
\alias{writeOutputFile}
\title{Generic function to write `data` to `outputFolder` with `fileName`}
\usage{
writeOutputFile(data, outputFolder, fileName)
}
\arguments{
\item{data}{tabular data}

\item{outputFolder}{str. Path to folder where `data` should be saved.}

\item{fileName}{str. Name of file to write in output.}
}
\description{
The data are saved in tab delimited format.
}
