% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveOutput.R
\name{writeProteinQuantIntensities}
\alias{writeProteinQuantIntensities}
\title{Save txt file with differential expression analysis results and the intensities used for the analysis.}
\usage{
writeProteinQuantIntensities(CompleteIntensityExperiment, outputFolder)
}
\arguments{
\item{CompleteIntensityExperiment}{Summarised experiment object returned by the `runGenericDiscovery` function. 
This contains the results of the differential expression analysis, the intensities used for the analysis and a masking logical matrix 
showing which intensities have been imputed.}

\item{outputFolder}{str. Path to folder where `data` should be saved.}
}
\description{
Save txt file with differential expression analysis results and the intensities used for the analysis.
}
