% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveOutputV2.R
\name{writeReplicateDataV2}
\alias{writeReplicateDataV2}
\title{Write protein_counts_and_intensity.json
This is a wrapper of `oneProteinReplDataGeneric`, looping over all proteins.}
\usage{
writeReplicateDataV2(longDTProt, outputFolder, GroupColumnName, GroupLabelType)
}
\arguments{
\item{longDTProt}{data.table protein information stored in long format.
Columns required: `ProteinId`, `GeneName`, `Description`, `log2NInt`, `Condition`,
 `Replicate`, `Imputed`.}

\item{outputFolder}{str. Path to folder where `data` should be saved.}
}
\description{
Write protein_counts_and_intensity.json
This is a wrapper of `oneProteinReplDataGeneric`, looping over all proteins.
}
