% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{compute_pcas}
\alias{compute_pcas}
\title{Compute PCAs}
\usage{
compute_pcas(
  Experiment,
  assayName,
  log,
  ndim = 2,
  scale_unit = FALSE,
  center = TRUE
)
}
\arguments{
\item{Experiment}{SummarizedExperiment object}

\item{assayName}{name of assay to use}

\item{ndim}{number of dimensions kept in result}

\item{scale_unit}{scale sample intensities to unit variance before computing PCAs}

\item{center}{center sample intensities means before computing PCAs}
}
\description{
Compute PCAs
}
