% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createCompleteIntensityExperiment}
\alias{createCompleteIntensityExperiment}
\title{Adds `intensityDF` information to `IntensityExperiment` with statistics about 
the number of replicates and number of imputed values in each condition of interest}
\usage{
createCompleteIntensityExperiment(IntensityExperiment, intensityDF)
}
\arguments{
\item{IntensityExperiment}{output from `constructSummarizedExperiment`}

\item{intensityDF}{Long format table with raw and imputed intensities. Each row is a feature (protein/peptide).}
}
\description{
Adds `intensityDF` information to `IntensityExperiment` with statistics about 
the number of replicates and number of imputed values in each condition of interest
}
