% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSummarizedExperiment.R
\name{createSummarizedExperiment}
\alias{createSummarizedExperiment}
\title{This function creates a summarized experiment object from a protein intensity table and experiment design.}
\usage{
createSummarizedExperiment(experimentDesign, proteinIntensities, listMetadata)
}
\arguments{
\item{experimentDesign}{data.frame. Experiment design provided in input by the user. Required columsn are: `SampleName` and `Condition`.}

\item{proteinIntensities}{data.frame. Wide matrix of intensities. Rows are proteins and columns are SampleNames. Required column: `ProteinId`.}

\item{listMetadata}{list of metadata: `Species`, `LabellingMethod`, `NormalisationAppliedToAssay`.}
}
\value{
A SummarizedExperiment object
}
\description{
This function creates a summarized experiment object from a protein intensity table and experiment design.
}
