% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaPairwiseOneCondition.R
\name{fitLinearModelLimma}
\alias{fitLinearModelLimma}
\title{Fit linear models using limma}
\usage{
fitLinearModelLimma(eset, designMat, contrastMatrix, metadataExperiment)
}
\description{
Fit linear models using limma
}
