% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaStatsFun.R
\name{limmaStatsFun}
\alias{limmaStatsFun}
\title{This function performs the differential expression analysis with limma including all pairwise comparisons 
using the condition provided}
\usage{
limmaStatsFun(
  ID_type,
  int_type,
  condition_col_name,
  run_id_col_name,
  rep_col_name,
  funDT,
  pairwise.comp = NULL,
  all.comparisons = TRUE,
  fix_distr = FALSE
)
}
\description{
This function performs the differential expression analysis with limma including all pairwise comparisons 
using the condition provided
}
