% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareEnrichment.R
\name{listComparisonExperiments}
\alias{listComparisonExperiments}
\title{This function creates a list of intensity experiments pertaining to each
comparison calculated by the limma pipeline}
\usage{
listComparisonExperiments(CompleteIntensityExperiment)
}
\description{
This function creates a list of intensity experiments pertaining to each
comparison calculated by the limma pipeline
}
