% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{mds_plot_experiment}
\alias{mds_plot_experiment}
\title{Plot first 2 dimensions of Multi-Dimensional Scaling plot}
\usage{
mds_plot_experiment(intensities, design, log = FALSE)
}
\arguments{
\item{intensities}{Matrix of intensities (rows are features, columns are samples)}

\item{design}{Experimental design}

\item{log}{logical}
}
\description{
Plot first 2 dimensions of Multi-Dimensional Scaling plot
}
