% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normaliseIntensity.R
\name{normaliseIntensity}
\alias{normaliseIntensity}
\title{This function normalises (based on `normalisationMethod`) the log2 intensities stored in the `log2NInt` column initialised by `initialiseLongIntensityDT`.}
\usage{
normaliseIntensity(longIntensityDT, normalisationMethod)
}
\arguments{
\item{longIntensityDT}{Output from `initialiseLongIntensityDT`}

\item{normalisationMethod}{logical. Whether to perform median normalisation by condition.}
}
\value{
log2NIntNorm If `normalisationMethod` is not `None` a new column is added containing the `log2NInt` normalised intensities
}
\description{
This function normalises (based on `normalisationMethod`) the log2 intensities stored in the `log2NInt` column initialised by `initialiseLongIntensityDT`.
}
\details{
Normalisation is only applied to non imputed intensities. `log2NIntNorm` is NA for imputed intensities.
}
