% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pd_iPRG_data}
\alias{pd_iPRG_data}
\title{Proteome Discoverer LFQ}
\format{
List containing the `intensities` table, the experiment `design` and the `parameters` following the Generic Format.
}
\usage{
pd_iPRG_data
}
\description{
Proteome Discoverer LFQ
}
\details{
The data can be used to run the entire workflow `runGenericDiscovery`. 
Download MaxQuant processed data from: https://app.massdynamics.com/experiments/1e2b17ed-835d-40bf-be75-c6a15f962ed0#/results-files
}
\references{
PRIDE Project PXD010981

Griss, Johannes, et al. 2019. “Spectral Clustering Improves Label-Free Quantification of Low-Abundant Proteins.” Journal of Proteome Research 18 (4): 1477–85.

Choi M,  et al. ABRF Proteome Informatics Research Group (iPRG) 2015 Study: Detection of Differentially Abundant Proteins in Label-Free Quantitative LC-MS/MS Experiments. J Proteome Res. 2017 16(2):945-957, PubMed: 27990823
}
\keyword{datasets}
