% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{plot_chosen_pca_experiment}
\alias{plot_chosen_pca_experiment}
\title{Plot principal components}
\usage{
plot_chosen_pca_experiment(
  Experiment,
  assayName = "intensities",
  dimPlot = c(1, 2),
  log = FALSE,
  auto_select_features = NULL,
  title = "PCA plot",
  format = "html"
)
}
\arguments{
\item{Experiment}{SummarizedExperiment object}

\item{assayName}{assay of intensities to use to compute PCs.}

\item{dimPlot}{vector of size 2 specifying the dimensions to plot.}

\item{log}{logical. Whether data should be logged before plotting}

\item{auto_select_features}{str. One of 'de' (differentially expressed features see Details), 
'hvf' (highly variable features). If not provided all features are kept.}

\item{title}{str. title on plot}

\item{format}{'pdf' or 'html'. Prepare image to be rendered for pdf or html Rmd output}
}
\description{
Plot principal components
}
\details{
#' A protein is defined DE if the adjusted PValue of the t-test or ANOVA (with multiple groups)
is less than 0.05.
}
