% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{plot_density_distr}
\alias{plot_density_distr}
\title{Density distribution of intensity values}
\usage{
plot_density_distr(Experiment, log = FALSE)
}
\arguments{
\item{Experiment}{SummarizedExperiment object}
}
\description{
Density distribution of intensity values
}
