% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{plot_imputed_vs_not}
\alias{plot_imputed_vs_not}
\title{Distribution of imputed versus non imputed values.}
\usage{
plot_imputed_vs_not(IntensityExperiment, CompleteIntensityExperiment)
}
\arguments{
\item{IntensityExperiment}{SummarizedExperiment object of raw data}

\item{CompleteIntensityExperiment}{SummarizedExperiment object of the data used for the DE analysis with limma}
}
\description{
Distribution of imputed versus non imputed values.
}
