% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{plot_replicate_missingness}
\alias{plot_replicate_missingness}
\title{Lollipop plot of missingness by intensity column}
\usage{
plot_replicate_missingness(Experiment)
}
\arguments{
\item{Experiment}{SummarizedExperiment object}
}
\description{
Lollipop plot of missingness by intensity column where missingness is defined as a value equal to 0 .
}
