% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareForFiltering.R
\name{prepareForFiltering}
\alias{prepareForFiltering}
\title{Filter protein IDs that are not present in at least one experiment}
\usage{
prepareForFiltering(funDT, condition_col_name, run_id_col_name)
}
\arguments{
\item{funDT}{data.table}

\item{condition_col_name}{name of column indicating the condition of interest}

\item{run_id_col_name}{name of column indicating the protein id}
}
\value{
`list`: filtered data.table and list of proteins
}
\description{
Filter protein IDs that are not present in at least one experiment
}
