% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{preparePlottingData}
\alias{preparePlottingData}
\title{Extract intensity and design from a SummarizedExperiment object}
\usage{
preparePlottingData(Experiment, log = FALSE)
}
\arguments{
\item{Experiment}{SummarizedExperiment object}

\item{log}{logical. Whether data should be logged before plotting}
}
\description{
Extract intensity and design from a SummarizedExperiment object
}
