% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcess.R
\name{preProcess}
\alias{preProcess}
\title{Prepare initial data for statistical modeling}
\usage{
preProcess(IntensityExperiment)
}
\arguments{
\item{IntensityExperiment}{SummarizedExperiment object as returned by `importData`}
}
\description{
In this steps condition are encoded to safe strings; intensities are normalised when required and missing
values imputed by MNAR
}
