% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{protein_missingness_experiment}
\alias{protein_missingness_experiment}
\title{Histogram of the distribution of missingness by protein where missingness is defined as 0 values.}
\usage{
protein_missingness_experiment(Experiment)
}
\arguments{
\item{Experiment}{SummarizedExperiment object}
}
\description{
Histogram of the distribution of missingness by protein where missingness is defined as 0 values.
}
