% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plotting_functions.R
\name{replicate_missingness_experiment}
\alias{replicate_missingness_experiment}
\title{this function creates a lollipop plot of missingness by intensity column where missingness is defined as a value equal to 0 .}
\usage{
replicate_missingness_experiment(intensities, design)
}
\arguments{
\item{intensities}{Matrix of intensities (rows are features, columns are samples)}

\item{design}{Experimental design}
}
\description{
this function creates a lollipop plot of missingness by intensity column where missingness is defined as a value equal to 0 .
}
