% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGenericDiscovery.R
\name{runGenericDiscovery}
\alias{runGenericDiscovery}
\title{This function orchestrates the MassExpression workflow (could be called by a  workflow step)}
\usage{
runGenericDiscovery(
  experimentDesign,
  proteinIntensities,
  normalisationMethod = "None",
  species,
  labellingMethod
)
}
\arguments{
\item{experimentDesign}{XX.}

\item{proteinIntensities}{XX}

\item{normalisationMethod}{Normalisation method. One of "none" or "median".}

\item{species}{Species. One of 'Human', 'Mouse', 'Yeast', 'Other'}

\item{labellingMethod}{One of 'LFQ' or 'TMT'}
}
\value{
List of two SummarisedExperiment objects: `IntensityExperiment` 
containing the raw intensities and  `CompleteIntensityExperiment` including 
imputed intensities and the results of the limma DE analysis.
}
\description{
This function orchestrates the MassExpression workflow (could be called by a  workflow step)
}
\examples{
design <- fragpipe_data$design
intensities <- fragpipe_data$intensities
parameters <- fragpipe_data$parameters
normalisation_method <- parameters[parameters[,1] == "UseNormalisationMethod",2]
species <- parameters[parameters[,1] == "Species",2]
labellingMethod <- parameters[parameters[,1] == "LabellingMethod",2]
listIntensityExperiments <- runGenericDiscovery(experimentDesign = design, 
proteinIntensities = intensities, 
normalisationMethod = normalisation_method,
species = species, 
labellingMethod = labellingMethod)
}
