% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLimmaPipeline.R
\name{runLimmaPipeline}
\alias{runLimmaPipeline}
\title{This function orchestrates imputation, normalization and the binary limma statistics accross all experimental comparisons}
\usage{
runLimmaPipeline(
  IntensityExperiment,
  normalisationMethod,
  fitSeparateModels,
  returnDecideTestColumn
)
}
\arguments{
\item{IntensityExperiment}{Output from createSummarizedExperiment}

\item{normalisationMethod}{logical. Whether to perform median normalisation by condition.}

\item{fitSeparateModels}{logical. TRUE to fit separate limma models for each pairwise comparisons 
(e.g. filtering and `lmFit` are run separately by comparison).}

\item{returnDecideTestColumn}{logical. If TRUE the row data of the `CompleteIntensityExperiment` will contain the output from 
`limma::decideTests`. 
if FALSE a single model is run for all contrasts.}
}
\description{
This function orchestrates imputation, normalization and the binary limma statistics accross all experimental comparisons
}
