% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLimmaPipeline.R
\name{runLimmaPipeline}
\alias{runLimmaPipeline}
\title{This function orchestrates imputation, normalization and the binary limma statistics accross all experimental comparisons}
\usage{
runLimmaPipeline(IntensityExperiment)
}
\arguments{
\item{experimentDesign}{Output from constructSummarizedExperiment}
}
\description{
This function orchestrates imputation, normalization and the binary limma statistics accross all experimental comparisons
}
