% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLimmaPipeline.R
\name{runLimmaPipeline}
\alias{runLimmaPipeline}
\title{This function orchestrates imputation, normalization and the binary limma statistics accross all experimental comparisons}
\usage{
runLimmaPipeline(IntensityExperiment, NormalisationMethod)
}
\arguments{
\item{NormalisationMethod}{logical. Whether to perform median normalisation by condition.}

\item{experimentDesign}{Output from constructSummarizedExperiment}
}
\description{
This function orchestrates imputation, normalization and the binary limma statistics accross all experimental comparisons
}
