% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveOutput.R
\name{saveOutput}
\alias{saveOutput}
\title{Save RData with SummarizedExperiment object and ProteinViz json}
\usage{
saveOutput(
  IntensityExperiment,
  CompleteIntensityExperiment,
  longIntensityDT,
  output_folder
)
}
\arguments{
\item{IntensityExperiment}{XX.}

\item{CompleteIntensityExperiment}{XX}

\item{longIntensityDT}{XX}

\item{output_folder}{XX}
}
\description{
Save RData with SummarizedExperiment object and ProteinViz json
}
