% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeProteinViz.R
\name{writeProteinViz}
\alias{writeProteinViz}
\title{This function writes a protein viz object which can be passed into MD's rails
database}
\usage{
writeProteinViz(outputFolder, IntensityExperiment)
}
\arguments{
\item{outputFolder}{Path to output folder}

\item{IntensityExperiment}{Output from runGenericDiscovery}
}
\description{
This function writes a protein viz object which can be passed into MD's rails
database
}
